% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A_crop_demand}
\alias{A_crop_demand}
\title{Demand for the crop (either nitrogen and phosphorus)}
\usage{
A_crop_demand(crop_abs, crop_exp_yield)
}
\arguments{
\item{crop_abs}{the absorption rate in [0,1] range. It can be looked up
as absorption coefficient in the Allegato 1 table and also by using \code{\link[=abs_n_coef_of]{abs_n_coef_of()}}
function for nitrogen or other similar functions for other nutrients.}

\item{crop_exp_yield}{the expected crop yield}
}
\value{
The crop demand in the same unit as \code{crop_exp_yield}, usually kg/ha.
}
\description{
N and P are estimated from expected yield and crop absorption rate
}
\details{
This is component A of the nitrogen and P2O5 fertilization plan balance
resulting as the crop expected yield times crop absorption rate.
}
\examples{
A_crop_demand(0.4, 1330)
}
