% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{abs_K_coef_of}
\alias{abs_K_coef_of}
\title{A function to get the Potassium absorption coefficient}
\usage{
abs_K_coef_of(crops, parts)
}
\arguments{
\item{crops}{a character vector of crop names to be looked up}

\item{parts}{a character vector of crop part names to be looked up. \code{R} recycling rules apply if number of parts is lower than number of crops}
}
\value{
a real number representing a percentage absorption of Potassium for the specific crop or \code{NA_real_} when no match is found or more than one match exists for the given \code{crop}, element and coefficient.
}
\description{
This closure function will use the absorption and removal table
to extract the K coefficient for \code{crops} and \code{parts} arguments
}
\examples{
# Returns 1
abs_K_coef_of("Ribes", "Pianta")
# Returns 0.0100 0.0851
abs_K_coef_of(c("Ribes", "Girasole"), "Pianta") / 100
}
