% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{abs_N_coef_of}
\alias{abs_N_coef_of}
\title{A function to get the Nitrogen absorption coefficient}
\usage{
abs_N_coef_of(crop)
}
\arguments{
\item{crop}{the crop to be looked up}
}
\value{
a real number representing a percentage absorption of Nitrogen for the specific crop or \code{NA_real_} when no match is found or more than one match exists for the given \code{crop}, element and coefficient.
}
\description{
function abs_N_coef_of
}
\details{
This closure function will use the subset Allegato 1 table
to perform a partial matching of its "crop" argument
}
\examples{
abs_N_coef_of("Ribes")                      # Returns 0.4
abs_N_coef_of(c("Ribes", "Girasole")) / 100 # Returns 0.0040 0.0431
}
