% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{abs_P_coef_of}
\alias{abs_P_coef_of}
\title{A function to get the Phosphorus absorption coefficient}
\usage{
abs_P_coef_of(crops, parts)
}
\arguments{
\item{crops}{a character vector of crop names to be looked up}

\item{parts}{a character vector of crop part names to be looked up. \code{R} recycling rules apply if number of parts is lower than number of crops}
}
\value{
a real number representing a percentage absorption of Phosphorus for the specific crop or \code{NA_real_} when no match is found or more than one match exists for the given \code{crop}, element and coefficient.
}
\description{
This closure function will use the absorption and removal table
to extract the N coefficient for \code{crops} and \code{parts} arguments
}
\examples{
# Returns 0.44
rem_K_coef_of("Ribes", "Frutti")
# Returns 0.0044 0.0115
rem_K_coef_of(c("Ribes", "Girasole"), "Frutti") / 100
# Returns 0.4
abs_P_coef_of("Ribes", "Pianta")
# Returns 0.0040 0.0019
abs_P_coef_of(c("Ribes", "Girasole"), "Pianta") / 100
}
