% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{b1_available_n}
\alias{b1_available_n}
\title{Estimate supply of Nitrogen in the soil for the crop (B1)}
\usage{
b1_available_n(total_n_pc, texture)
}
\arguments{
\item{total_n_pc}{Total Nitrogen percentage in soil as appropriatly sampled}

\item{texture}{Soil texture from table 1 (one of "Sandy", "Loam", "Clayey") as appropriatly sampled}
}
\value{
The available N in soil in kg/ha
}
\description{
Estimate supply of Nitrogen in the soil for the crop (B1)
}
\examples{
b1_available_n(0.139, "Clayey")                    # Returns 3.3777 kg/ha
b1_available_n(c(0.139, 0.5), c("Clayey", "Loam")) # Returns  3.3777 13.0000
}
