% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{b1_available_n_for}
\alias{b1_available_n_for}
\title{Coefficient of supply of Nitrogen for a soil texture in table 1}
\usage{
b1_available_n_for(texture)
}
\arguments{
\item{texture}{Soil texture (one of "Sandy", "Loam", "Clayey", Guidelines ed. year 2020 page 21, table 2)}
}
\value{
Nitrogen coefficient
}
\description{
The Nitrogen coefficient has to be applied to the total Nitrogen soil content
percentage to get soil available Nitrogen (B1)
}
\examples{
\dontrun{ b1_available_n_for("Sandy") # Returns 28.4 }
}
