% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{b2_mineralized_n_coeff_for}
\alias{b2_mineralized_n_coeff_for}
\title{Coefficients of Nitrogen mineralization in soil by CN ratios and soil textures}
\usage{
b2_mineralized_n_coeff_for(cn_ratio, texture)
}
\arguments{
\item{cn_ratio}{Carbon Nitrogen ratio in soil}

\item{texture}{Soil texture (one of "Sandy", "Loam", "Clayey", table 2, page 23 of the 'Disciplinare')}
}
\value{
a vector of Nitrogen mineralization coefficients
}
\description{
Coefficients of Nitrogen mineralization in soil by CN ratios and soil textures
}
