% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{B_N_in_soil}
\alias{B_N_in_soil}
\title{Supply of Nitrogen from soil fertility}
\usage{
B_N_in_soil(b1, b2)
}
\arguments{
\item{b1}{Available Nitrogen in soil, typically from function \code{b1_available_n}}

\item{b2}{Mineralized Nitrogen from SOM, typically from function \code{b2_mineralized_n}}
}
\value{
Total Nitrogen in soil as the sum of available and mineralized Nitrogen
        Note that N supply is multipled by -1 before being returned because B has to
        be subtracted to the total N fertilization!
}
\description{
This is component B of the fertilization plan equation, resulting from the available nitrogen in soil and nitrogen mineralized from soil organic matter
}
