% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{c1_coeff_leached_n_precip}
\alias{c1_coeff_leached_n_precip}
\title{Percentage of Nitrogen leached by precipitation}
\usage{
c1_coeff_leached_n_precip(rainfall_oct_jan)
}
\arguments{
\item{rainfall_oct_jan}{cumulative precipitation in mm in the 4 months-period October - January}
}
\value{
Nitrogen coefficient leached [0..1]. This coefficient should be applied to the available Nitrogen
}
\description{
Estimates the ratio of Nitrogen leached by using cumulative precipitation in the period
October 1st - January 31st as described on pages 24 and 25 of the "Disciplinare".
}
\examples{
# Returns 0.00 0.25 1.00
\dontrun{ c1_coeff_leached_n_precip(c(30, 175, 270))  }
}
