% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C_N_drain_leach}
\alias{C_N_drain_leach}
\title{Loss of Nitrogen leached yearly by soil drainage rate and texture}
\usage{
C_N_drain_leach(drainage_rate, soil_texture)
}
\arguments{
\item{drainage_rate}{Rate of drainage in soil (either "fast", "normal", "slow", "no drainage")}

\item{soil_texture}{Soil texture (either "Sandy", "Loam", or "Clayey")}
}
\value{
Nitrogen leaching from soil in kg/ha/y
}
\description{
Rate of Nitrogren in kg per hactare per year, as in table 3, page 25 of the 'Disciplinare'.
This is one of two mutually exclusive methods (c1 and c2) to compute Nitrogen leaching,
the other being c1 "Metodo in base alle precipitazioni"
}
\examples{
# Returns 30 50
C_N_drain_leach(c("Rapido", "Lento"), c("Argilloso", "Sabbioso"))
}
