% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C_N_precip_leach}
\alias{C_N_precip_leach}
\title{Loss of Nitrogen leached by precipitation}
\usage{
C_N_precip_leach(available_n, rainfall_oct_jan)
}
\arguments{
\item{available_n}{available Nitrogen for the crop, usually returned by \code{\link{b1_available_n}}}

\item{rainfall_oct_jan}{cumulative precipitation in mm in the 4 months-period October - January}
}
\value{
Nitrogen leaching from soil in the year the precipitation figure is collected, in kg/ha
}
\description{
Estimates Nitrogen leached by using cumulative precipitation in the period
October 1st - January 31st as described on pages 24 and 25 of the "Disciplinare".
This is one of two mutually exclusive methods (c1 and c2) to compute Nitrogen leaching,
the other being c1 "Metodo in base alla facilità di drenaggio"
The leaching affects only the available Nitrogen part (not total Nitrogen)
}
\examples{
# Returns 3.3777 i.e. all available Nitrogen was leached
C_N_precip_leach(3.3777, 350)
}
