% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C_P_immob_by_Ca}
\alias{C_P_immob_by_Ca}
\title{Loss of Phosphorus due to immobilization from Limestone content}
\usage{
C_P_immob_by_Ca(Ca_pc, soil_texture)
}
\arguments{
\item{Ca_pc}{Limestone soil content in \% of Cation-exchange capacity}

\item{soil_texture}{Soil texture (either "Sandy", "Loam", or "Clayey")}
}
\value{
Phosphorus (P2O5) to be supplied (positive sign) to soil in kg/ha. This is a
correction factor that takes into account the unavailable P quantity due to limestone content
}
\description{
Supply of Phosphorus (P2O5) in kg per hactare to counteract immobilization due to limestone
}
\examples{
# Returns 3.246
C_P_immob_by_Ca(92.3, "Argilloso")
}
