% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{coef_maker}
\alias{coef_maker}
\title{Look for a crop fertilization absorption or removal of a specific fertilizer}
\usage{
coef_maker(abs_or_removal, nutrient)
}
\arguments{
\item{abs_or_removal}{Either "ass." for absorption or "ass." for removal}

\item{nutrient}{Either "N", "P2OS", or "K2O"}
}
\value{
function
}
\description{
The function does not perform any actual lookup but merely returns a function
that performs the search on a subset of the Allegato 1 table.
}
