% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{crop_type_lookup}
\alias{crop_type_lookup}
\title{Time coefficient for organic matter mineralization}
\usage{
crop_type_lookup(crop_type)
}
\arguments{
\item{crop_type}{Crop type for estimation of the time coefficient (Attachment 2 table and page 29 of the 'Disciplinare')}
}
\value{
The time coefficient
}
\description{
Used internally by \code{\link{b2_mineralized_n}}
}
