% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.R
\name{D_N_denitrification}
\alias{D_N_denitrification}
\title{Loss of Nitrogen with denitrification, adsorbation, volatilization}
\usage{
D_N_denitrification(B, drainage_rate, soil_texture)
}
\arguments{
\item{B}{Nitrogen supply from soil fertility estimate, usually derived with function \code{\link{B_N_in_soil}}}

\item{drainage_rate}{Rate of drainage in soil (either "fast", "normal", "slow", "no drainage")}

\item{soil_texture}{Soil texture (either "Sandy", "Loam", or "Clayey")}
}
\value{
Nitrogen denitrification coefficient
}
\description{
Coefficient of Nitrogen denitrification, adsorbation, volatilization processes in soil,
as found in page 25 of the "Disciplinare".
}
\examples{
\dontrun{ D_N_denitrification(30.98, "slow", "Clayey")  # Returns 13.941 }
}
