% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E_N_from_prev_crop}
\alias{E_N_from_prev_crop}
\title{Supply of Nitrogen quantity in soil from previous crop}
\usage{
E_N_from_prev_crop(crop)
}
\arguments{
\item{crop}{Previous crop as defined in table 05 of the "Disciplinare", page 26}
}
\value{
Nitrogen quantity in kg/ha, either positive (less N available) or negative (increased N available).
        Note that N quantity is multipled by -1 before being returned because E has to be subtracted to the total N fertilization!
}
\description{
This is a simple table lookup (table 05 of the "Disciplinare"). Partial matching is respected so that
\code{E_n_from_prev_crop("Gira")} is matched to "Girasole" and the function returns 0 but
\code{E_n_from_prev_crop("Prati")} does not match a unique record so that the function returns NA with a warning.
}
\details{
Table 05 is a \code{data.table} with values for Nitrogen quantity in kg/ha of `r nrows(tables_l$tab_05_dt)` typical crops:
`r tables_l$tab_05_dt`
}
\examples{
# Returns 0 kg/ha
E_N_from_prev_crop("Girasole")
# Returns 0 10 kg/ha
E_N_from_prev_crop(c("Girasole", "Mais: stocchi asp"))
}
