% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{F_K_in_soil}
\alias{F_K_in_soil}
\title{Supply of potassium from soil fertility}
\usage{
F_K_in_soil(k_ppm, soil_texture, soil_depth_cm)
}
\arguments{
\item{k_ppm}{Current potassium concentration in soil (in ppm or mg/kg)}

\item{soil_texture}{Soil texture (one of \code{Argilloso}, \code{Franco}, \code{Sabbioso})}

\item{soil_depth_cm}{Depth of soil tillage in cm (usually 30 or 40 cm)}
}
\value{
Total potassium (K2O) quantity in excess (negative sign) or in demand (positive sign, hence to be supplied)
due to its fertility
}
\description{
This is component F of the potassium fertilization plan balance.
}
\note{
For soil depth values outside the typical range [30,40] cm a warning is issued.
}
\examples{
# Returns -976.47
F_K_in_soil(449, "Argilloso", 30)
}
