% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{F_N_prev_fertilization}
\alias{F_N_prev_fertilization}
\title{Supply of Nitrogen from previous organic fertilizatons}
\usage{
F_N_prev_fertilization(n_supply, organic_fertilizer, years_ago)
}
\arguments{
\item{n_supply}{Supply of nitrogen in kg/ha from the organic fertilization performed}

\item{organic_fertilizer}{Type of organic fertilizer as found in table 06 (one of \code{Ammendanti}, \verb{Liquame bovino}, \verb{Liquame suino e pollina})}

\item{years_ago}{Numeric, the number of years passed from last performed fertilization}
}
\value{
Estimate of supply of nitrogen currently still in soil
}
\description{
Supply of Nitrogen from previous organic fertilizatons
}
