% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertplan-package.R
\docType{package}
\name{fertplan}
\alias{fertplan}
\title{fertplan}
\description{
fertplan is a small R package to be used in aiding the computation of
fertilization plans for erbaceous and perennial crops.
fertplan computes nitrogen, phosphorus (P2O5), and potassium (K2O)
concentrations to be supplied to field crops.
}
\details{
It strictly follows the regulation for integrated crop production
published by the administrative authority of Lazio region, Italy.
Assessorato Agricoltura, Promozione della Filiera e della Cultura del
Cibo, Ambiente e Risorse Naturali, 2020. Parte agronomica, norme generali,
Disciplinare di Produzione Integrata della Regione Lazio - SQNPI.
Regione Lazio, Roma, Italy.
\url{http://www.regione.lazio.it/rl_agricoltura/?vw=documentazioneDettaglio&id=52065}
}
\author{
Marco Bascietto <marco.bascietto@crea.gov.it>
}
