% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\name{G_K_immob_by_clay}
\alias{G_K_immob_by_clay}
\title{Loss of Potassium due to immobilization from Limestone content}
\usage{
G_K_immob_by_clay(clay_pc)
}
\arguments{
\item{clay_pc}{Clay content in soil in \%}
}
\value{
Potassium (K2O) to be supplied (positive sign) to soil in kg/ha. This is a
correction factor that takes into account the unavailable K quantity due to clay content
}
\description{
Supply of Potassium (K2O) in kg per hactare to counteract immobilization due to Clay
}
\examples{
# Returns 1.72 kg/ha
G_K_immob_by_clay(40)
}
