% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\name{G_N_from_atmosphere}
\alias{G_N_from_atmosphere}
\title{Supply of nitrogen from atmosphere or from nitrogen-fixing bacteria}
\usage{
G_N_from_atmosphere(coeff)
}
\arguments{
\item{coeff}{a simple ratio [0,1] to linearly correct the
estimate, 1 to estimate 20 kg/ha nitrogen, 0 to estimate 0 kg/ha.}
}
\value{
Estimate of N from atmosphere of from bacteria in kg/ha
}
\description{
Yearly availability is estimated to be 20 kg/ha in levelled crops close
to urban settlements. This figure has to be appropriately adapted
to each crop through a [0,1] coefficient. Page 25 of 2020 Guidelines.
Note that the N estimate is given in negative sign (ie a flow into the soil).
}
\examples{
# Returns -10 kg/ha
G_N_from_atmosphere(0.5)
}
