% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{get_matching_values}
\alias{get_matching_values}
\title{Match elements among a vector, allowing for duplicates, return elements from a third vector}
\usage{
get_matching_values(x_v, lookup_v, match_v)
}
\arguments{
\item{x_v}{the values to be matched}

\item{lookup_v}{the values to be matched against}

\item{match_v}{the values to be returned by index}
}
\value{
avector of same length as \code{x}
}
\description{
Internal function, uses internally \code{pmatch}.
}
\examples{
\dontrun{# Returns 10 20 10
get_matching_values(
  c(1,2,1),
  c(3,6,3,7,7,3,2,1,1),
  c(30,60,30,70,70,30,20,10,10))}
}
