% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_formula_f.R
\name{k_immobilization}
\alias{k_immobilization}
\title{Internal function to compute K immobilization}
\usage{
k_immobilization(clay_pc)
}
\arguments{
\item{clay_pc}{Clay content in soil (\%)}
}
\value{
Immobilization factor for Potassium
}
\description{
No check on function parameters is performed, this is left to
the package function calling k_immobilization
For this reason this function is not exported by package.
}
\details{
Ref. Disciplinare, ed. 2020, page 34
}
