% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_f.R
\name{leached_n_coeff}
\alias{leached_n_coeff}
\title{Internal function to compute coefficient of N leaching}
\usage{
leached_n_coeff(rainfall)
}
\arguments{
\item{rainfall}{Cumulative rainfall between october and january in mm}
}
\value{
An empirical estimate of the ration of Nitrogen leached due to rainfall
}
\description{
This is method c1 (Nitrogen loss due to Rainfall leaching effect). This is described on page 22 of the 2020 guidelines.
}
