% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{lookup_var_by_crop_texture}
\alias{lookup_var_by_crop_texture}
\title{Matches soil texture and crop features to dt}
\usage{
lookup_var_by_crop_texture(dt, crop, soil_texture)
}
\arguments{
\item{dt}{A table from the "Disciplinare" featuring soil_texture and crop columns}

\item{crop}{Crop character vector describing the crop}

\item{soil_texture}{Soil texture (either "Sandy", "Loam", or "Clayey")}
}
\value{
A \code{data.table} matching \code{dt} by \code{crop} and \code{soil_texture}
}
\description{
Matches soil texture and crop features to dt
}
