% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{lookup_var_by_drainage_texture}
\alias{lookup_var_by_drainage_texture}
\title{Matches drainage rate and texture features to dt}
\usage{
lookup_var_by_drainage_texture(dt, drainage_rate, texture)
}
\arguments{
\item{dt}{A table from the "Disciplinare" featuring drainage and soil_texture columns}

\item{drainage_rate}{Rate of drainage in soil (either "fast", "normal", "slow", "no drainage")}

\item{texture}{Soil texture (either "Sandy", "Loam", or "Clayey")}
}
\value{
A \code{data.table} matching \code{dt} by \code{drainage_rate} and \code{texture}
}
\description{
Matches drainage rate and texture features to dt
}
\examples{
# Returns a data.table:
#    drainage soil_texture n_denitrificated_coeff
# 1:     slow       Clayey                   0.45
\dontrun{lookup_var_by_drainage_texture(tables_l[["tab_04_dt"]], "slow", "Clayey")}
# Returna a data.table:
#    drainage soil_texture n_leached_kg_ha_y
# 1:     slow       Clayey                50
\dontrun{lookup_var_by_drainage_texture(tables_l[["tab_03_dt"]], "slow", "Clayey")}
}
