% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{lookup_var_by_fertilizer_year}
\alias{lookup_var_by_fertilizer_year}
\title{Matches fertilizer and frequency features to dt}
\usage{
lookup_var_by_fertilizer_year(dt, fertilizer, years)
}
\arguments{
\item{dt}{A table from the "Disciplinare" featuring organic_fertilizer and frequency columns}

\item{fertilizer}{Character vector describing type organic fertilizer}

\item{years}{Frequency of fertilization}
}
\value{
A \code{data.table} matching \code{dt} by \code{organic_fertilizer} and \code{frequency}
}
\description{
Matches fertilizer and frequency features to dt
}
