% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{mineralized_N_coeff_from}
\alias{mineralized_N_coeff_from}
\title{Supply of Nitrogen mineralization in soil (coefficient)}
\usage{
mineralized_N_coeff_from(cn_ratio, texture)
}
\arguments{
\item{cn_ratio}{Carbon Nitrogen ratio in soil}

\item{texture}{Soil texture (one of "Sandy", "Loam", "Clayey", Guidelines ed. year 2020 page 21, table 2)}
}
\value{
The coefficient of Nitrogen mineralization
}
\description{
Lookup in table 2 for the coefficient of Nitrogen mineralization in soil
}
