% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_f.R
\name{natural_n}
\alias{natural_n}
\title{Internal function to get N from atmospehere or n-fixing bacteria}
\usage{
natural_n(coeff)
}
\arguments{
\item{coeff}{a simple ratio [0..1] to linearly correct the
estimate, 1 to estimate 20 kg/ha nitrogen, 0 to estimate 0 kg/ha.}
}
\value{
Estimate of N from atmosphere of from bacteria in kg/ha
}
\description{
Page 25 of 2020 Guidelines. Estimate is given in negative sign (ie a flow into the soil)
}
