% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_f.R
\name{p_immobilization}
\alias{p_immobilization}
\title{Internal function to compute P immobilization}
\usage{
p_immobilization(a_coeff, Ca_pc)
}
\arguments{
\item{a_coeff}{a coefficient, depends on soil texture}

\item{Ca_pc}{Limestone content in soil (\%)}
}
\value{
Immobilization factor for Phosphorus
}
\description{
No check on function parameters is performed, this is left to
the package function calling p_immobilization
For this reason this function is not exported by package.
}
\details{
Ref. Disciplinare, ed. 2020, page 32
}
