% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_f.R
\name{pk_availability}
\alias{pk_availability}
\title{Internal function to compute P and K availability}
\usage{
pk_availability(soil_depth_cm, Da, Q)
}
\arguments{
\item{soil_depth_cm}{Soil depth in cm. This is the P constant in
'Disciplinare' equation, eg soil depth / 10}

\item{Da}{Apparent density, depends on soil texture}

\item{Q}{Delta between "normal" element (P or K)
availability in soil and actual one}
}
\value{
Nutrient (P or K) supply to the soil (positive sign) or excess (negative sign)
}
\description{
No check on function parameters is performed, this is left to
the package function calling p_immobilization
For this reason this function is not exported by package.
}
\details{
Ref. Disciplinare, ed. 2020, page 31, 34
}
