% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{abs_K_coef_of}
\alias{abs_K_coef_of}
\title{A function to get the Potassium absorption coefficient}
\usage{
abs_K_coef_of(crop)
}
\arguments{
\item{crop}{the crop to be looked up}
}
\value{
a real number representing a percentage absorption of Potassium (K2O) for the specific crop or \code{NA_real_} when no match is found or more than one match exists for the given \code{crop}, element and coefficient.
}
\description{
function abs_K_coef_of
}
\details{
This closure function will use the subset Allegato 1 table
to perform a partial matching of its "crop" argument
}
\examples{
# Returns 1
abs_K_coef_of("Ribes")
# Returns 0.0100 0.0851
abs_K_coef_of(c("Ribes", "Girasole")) / 100
}
