% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{b1_available_n}
\alias{b1_available_n}
\title{Estimate supply of Nitrogen in the soil for the crop (B1)}
\usage{
b1_available_n(total_n_pc, texture)
}
\arguments{
\item{total_n_pc}{Total nitrogen percentage in soil}

\item{texture}{Soil texture (one of \code{Argilloso}, \code{Franco}, \code{Sabbioso})}
}
\value{
The available N in soil in kg/ha
}
\description{
Estimate supply of Nitrogen in the soil for the crop (B1)
}
\examples{
b1_available_n(0.139, "Argilloso") # Returns 3.3777 kg/ha
b1_available_n(c(0.139, 0.5), c("Argilloso", "Franco")) # Returns  3.3777 13.0000
}
