% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{b2_mineralized_n}
\alias{b2_mineralized_n}
\title{Supply of Nitrogen mineralized from soil organic matter}
\usage{
b2_mineralized_n(crop_type, som_pc, cn_ratio, texture)
}
\arguments{
\item{crop_type}{Crop type for estimation of the time coefficient (Guidelines ed. year 2020 page 22 and Table 15.3 page 67)}

\item{som_pc}{Soil Organic Matter percentage}

\item{cn_ratio}{Carbon Nitrogen ratio in soil}

\item{texture}{Soil texture (one of "Sandy", "Loam", "Clayey", Guidelines ed. year 2020 page 21, table 2)}
}
\value{
Quantity of Nitrogen in kg/ha
}
\description{
Note that values of soil organic matter >3% are kept constant at 3%
}
\examples{
# Returns 20.7 kg/ha
b2_mineralized_n("Girasole", 2.3, 9.57, "Clayey")
}
