% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{B_N_in_soil}
\alias{B_N_in_soil}
\title{Supply of Nitrogen from soil fertility}
\usage{
B_N_in_soil(b1, b2)
}
\arguments{
\item{b1}{Available nitrogen in soil, typically from function \code{\link[=b1_available_n]{b1_available_n()}}}

\item{b2}{Mineralized nitrogen from SOM, typically from function \code{\link[=b2_mineralized_n]{b2_mineralized_n()}}}
}
\value{
Total nitrogen in soil as the sum of available and mineralized Nitrogen
Note that N supply is multipled by -1 before being returned because B has to
be subtracted to the total N fertilization!
}
\description{
This is component B of the fertilization plan balance, resulting from the available nitrogen in soil and nitrogen mineralized from soil organic matter
}
