% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{crop_type_lookup}
\alias{crop_type_lookup}
\title{Time coefficient for organic matter mineralization}
\usage{
crop_type_lookup(crop_type)
}
\arguments{
\item{crop_type}{Crop type for estimation of the time coefficient (Guidelines ed. year 2020 page 22 and Table 15.3 page 67)}
}
\value{
The time coefficient
}
\description{
Used internally by \code{\link{b2_mineralized_n}}
}
