% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{F_K_in_soil}
\alias{F_K_in_soil}
\title{Supply of Potassium from soil fertility}
\usage{
F_K_in_soil(k_ppm, soil_texture, soil_depth_cm)
}
\arguments{
\item{k_ppm}{Current Potassium concentration in soil (in ppm or mg/kg)}

\item{soil_texture}{Soil texture (one of "Sandy", "Loam", "Clayey", table 2, page 23 of the 'Disciplinare')}

\item{soil_depth_cm}{Depth of soil tillage in cm (usually 30 or 40 cm)}
}
\value{
Total Potassium (K2O) quantity in excess (negative sign) or in demand (positive sign, hence to be supplied)
                       due to its fertility
}
\description{
This is component F of the Potassium fertilization plan equation
}
\examples{
# Returns -976.47
F_K_in_soil(449, "Clayey", 30)
}
