% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{F_N_prev_fertilization}
\alias{F_N_prev_fertilization}
\title{Supply of Nitrogen from previous organic fertilizatons}
\usage{
F_N_prev_fertilization(n_supply, organic_fertilizer, years_ago)
}
\arguments{
\item{n_supply}{Supply of Nitrogen in kg/ha from the organic fertilization performed}

\item{organic_fertilizer}{Type of organic fertilizer as found in table 06 ("Bovine manure", "Poultry manure", "Conditioners")}

\item{years_ago}{Numeric, the number of years passed from last perfored fertilization}
}
\value{
Estimate of supply of Nitrogen currently still in soil
}
\description{
Supply of Nitrogen from previous organic fertilizatons
}
