% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{get_available}
\alias{get_available}
\title{Get available crops localized in current language}
\usage{
get_available(variable)
}
\value{
a character vector of available crops
}
\description{
Get available crops localized in current language
}
\examples{
get_available("crop")
}
