% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_f.R
\name{get_available}
\alias{get_available}
\title{Get available variable values localized in current language}
\usage{
get_available(variable = NULL, aux = NULL, is_removal = FALSE)
}
\arguments{
\item{variable}{character vector pointing at the tabled variable to look up.
One of \code{crop}, \verb{crop by group}, \verb{crop by part group}, \verb{crop type}, \code{part}, \code{drainage}, \verb{soil texture}, \verb{organic fertilizer}, \code{frequency}, \verb{previous crops}, \verb{crop class}, \code{languages} or \code{NULL} (default).}

\item{aux}{character vector useful to pass the part when variable is "crop by part group"}

\item{is_removal}{boolean value useful to select only crops featuring removal
coefficients, for variable assuming values "crop", "crop by group", "crop by part group", and "part"}
}
\value{
a character vector of available tabled variable values, a list of
variable values (when \code{variable} is "group_crop"), or a character vector
of available variables to be looked up (when \code{variable} is \code{NULL}),
or \code{NULL} when no matching variables were passed.
}
\description{
Get available variable values localized in current language
}
\examples{
head(get_available("crop"))
get_available("languages")
}
