% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H_K_leaching}
\alias{H_K_leaching}
\title{Potassium leaching due to clay content in soil}
\usage{
H_K_leaching(clay_pc)
}
\arguments{
\item{clay_pc}{Clay content in soil in \%}
}
\value{
Supply of K2O in kg/ha
}
\description{
K leaching is estimated from the amount of clay in soil in percentage
(the higher clay amount the lower K is leached)
}
\examples{
# Returns 60 60 60 10 30 kg/ha
H_K_leaching(c(0, 4, 5, 26, 13))
}
