% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n_switch.R
\name{i18n_switch}
\alias{i18n_switch}
\title{Switch language}
\usage{
i18n_switch(lang)
}
\arguments{
\item{lang}{Character vector, one of "it", "en"}
}
\description{
The default guidelines are in italian language. Switching to english or other available languages aids in
selecting the appropriate variables values.
}
\examples{
get_available("crop")[1:10]
i18n_switch("en")
get_available("crop")[1:10]
}
