% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{rem_N_coef_of}
\alias{rem_N_coef_of}
\title{A function to get the Nitrogen removal coefficient}
\usage{
rem_N_coef_of(crops, parts)
}
\arguments{
\item{crops}{a character vector of crop names to be looked up}

\item{parts}{a character vector of crop part names to be looked up. \code{R} recycling rules apply if number of parts is lower than number of crops}
}
\value{
a real number representing a percentage removal of Nitrogen for the specific crop or \code{NA_real_} when no match is found or more than one match exists for the given \code{crop}, element and coefficient.
}
\description{
This closure function will use the absorption and removal table
to extract the N coefficient for \code{crops} and \code{parts} arguments
}
\examples{
rem_N_coef_of("Ribes", "Frutti")                      # Returns 0.14
rem_N_coef_of(c("Ribes", "Girasole"), "Frutti") / 100 # Returns 0.0014 0.0280
}
